update the cache without `gm cache --update`
when cloning or removing a project, there is no need to update the
whole cache, that takes unnecessary time...

this commit only appends the new project to the cache or removes it
when using `gm clone` or `gm remove`.
because it started to be used in multiple places, i've refactored
the check of the cache existence in `check-cache-file` in the
`fs/store/` module.