test imports exactly (#126)
this PR ensures that the import tests are _strong_.

by _strong_, i mean that now, they compare the list of commands in scope
before and after the `use` of a given module.
the diff, i.e. the new commands in scope, should be exactly the commands
from the module.
the import tests don't rely on a prefix anymore.